+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.1.7 - 2021-04-29
+### Added
+- Added Invoice variable to twig
+
+### Changed
+- Changed query to include sent paramater
+
 ## 0.1.6 - 2021-04-27
 - Automatically mark invoices as sent.